chore: Replace Property with DataBrokerRequests
The "Property" name was not a good reflection of the combination of
the variables vssPath + Field (grpc). This model is not replaced with
explicit *Request classes.

BREAKING CHANGE: The API for fetch / subscribe / update did change.
The input for these methods are now wrapped via "DataBrokerRequest"
classes e.g. "FetchRequest", "UpdateRequest". For generated VSS models
the classes "VssNodeFetchRequest" etc. should be used.

- The Property model was removed in favor of DataBrokerRequest