fix: Fix UInt not supported for VssNodes
UInt could not be used before this commit because @JVMOverloads was
used for all generated VSS data classes. Since UInt is a Kotlin
specific data type it was incompatible with the annotation. As long as
we are only using one constructor param for VssSignals we can remove
the annotation because the default constructor will still be
available.

Note: The java byte code will still show an Int for the Kotlin UInt
model. This is fine as long as the SDK is speaking Kotlin while
interacting with these models.

close #93